feat: skill to eo lo

+++ Shared/Dtos/PlayerDto.cs
@@ -4,5 +4,5 @@ public class PlayerDto
 {
     public string NickName { get; set; }
     public Uri? ImageUrl { get; set; }
-    public int Elo { get; set; }
+    public decimal Elo { get; set; }
 }
